Extract parseTarget to utility function
- Move parseTarget method from LaunchQLProject class to utils/target-utils.ts
- Update imports in launchql.ts and target-api.test.ts
- Maintain all existing functionality and test coverage
- Improve code organization by separating utility functions from class methods
