+++ packages/core/src/core/class/launchql.ts
@@ -34,6 +34,7 @@ import {
 } from '../../modules/modules';
 import { packageModule } from '../../packaging/package';
 import { extDeps, resolveDependencies } from '../../resolution/deps';
+import { parseTarget } from '../../utils/target-utils';
 
 
 const logger = new Logger('launchql');
@@ -646,62 +647,6 @@ export class LaunchQLProject {
 
   // ──────────────── Project Operations ────────────────
 
-  private parseTarget(target: string): { projectName: string; toChange?: string } {
-    if (!target) {
-      throw new Error('Target parameter is required');
-    }
-
-    if (target.includes(':@')) {
-      const atIndex = target.indexOf(':@');
-      const beforeAt = target.substring(0, atIndex);
-      const afterAt = target.substring(atIndex + 2);
-      
-      if (!afterAt) {
-        throw new Error(`Invalid tag format: ${target}. Expected format: project:@tagName`);
-      }
-      
-      // Check if this is a simple project:@tag format
-      if (!beforeAt.includes(':')) {
-        if (!beforeAt) {
-          throw new Error(`Invalid tag format: ${target}. Expected format: project:@tagName`);
-        }
-        return { projectName: beforeAt, toChange: `${beforeAt}:@${afterAt}` };
-      }
-      
-      const lastColonIndex = beforeAt.lastIndexOf(':');
-      const projectName = beforeAt.substring(0, lastColonIndex);
-      const changeName = beforeAt.substring(lastColonIndex + 1);
-      
-      if (!projectName || !changeName) {
-        throw new Error(`Invalid tag format: ${target}. Expected format: project:@tagName`);
-      }
-      
-      const toChange = `${changeName}:@${afterAt}`;
-      return { projectName, toChange };
-    }
-    
-    if (target.includes(':') && !target.includes('@')) {
-      const parts = target.split(':');
-      
-      if (parts.length > 2) {
-        throw new Error(`Invalid target format: ${target}. Expected formats: project, project:changeName, or project:@tagName`);
-      }
-      
-      const [projectName, changeName] = parts;
-      
-      if (!projectName || !changeName) {
-        throw new Error(`Invalid change format: ${target}. Expected format: project:changeName`);
-      }
-      
-      return { projectName, toChange: changeName };
-    }
-    
-    if (!target.includes(':')) {
-      return { projectName: target, toChange: undefined };
-    }
-
-    throw new Error(`Invalid target format: ${target}. Expected formats: project, project:changeName, or project:@tagName`);
-  }
 
   async deploy(
     opts: LaunchQLOptions,
@@ -724,7 +669,7 @@ export class LaunchQLProject {
         throw new Error('Not in a LaunchQL workspace or module');
       }
     } else {
-      const parsed = this.parseTarget(target);
+      const parsed = parseTarget(target);
       name = parsed.projectName;
       toChange = parsed.toChange;
     }
@@ -878,7 +823,7 @@ export class LaunchQLProject {
         throw new Error('Not in a LaunchQL workspace or module');
       }
     } else {
-      const parsed = this.parseTarget(target);
+      const parsed = parseTarget(target);
       name = parsed.projectName;
       toChange = parsed.toChange;
     }
@@ -981,7 +926,7 @@ export class LaunchQLProject {
         throw new Error('Not in a LaunchQL workspace or module');
       }
     } else {
-      const parsed = this.parseTarget(target);
+      const parsed = parseTarget(target);
       name = parsed.projectName;
       toChange = parsed.toChange;
     }
